requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-dotenv==1.0.0
mistralai==0.0.12
python-multipart==0.0.9
httpx==0.25.2
gunicorn==21.2.0